JSSE14SocketFactory merged to JSSESocketFactory
- Shortened the inheritance to GlassfishServerSocketFactory
