+++ nucleus/security/core/src/main/java/com/sun/enterprise/security/ssl/GlassfishServerSocketFactory.java
@@ -18,20 +18,23 @@
 package com.sun.enterprise.security.ssl;
 
 import java.io.IOException;
+import java.lang.System.Logger;
+import java.lang.System.Logger.Level;
 import java.security.KeyStore;
 import java.util.Objects;
-import java.util.logging.Level;
 
 import javax.net.ssl.KeyManager;
 import javax.net.ssl.X509KeyManager;
 
-import org.glassfish.grizzly.config.ssl.JSSE14SocketFactory;
+import org.glassfish.grizzly.config.ssl.JSSESocketFactory;
 import org.glassfish.hk2.api.ServiceLocator;
 
 /**
  * @author Sudarsan Sridhar
  */
-public class GlassfishServerSocketFactory extends JSSE14SocketFactory {
+public class GlassfishServerSocketFactory extends JSSESocketFactory {
+
+    private static final Logger LOG = System.getLogger(GlassfishServerSocketFactory.class.getName());
 
     private final ServiceLocator locator;
     private SSLUtils sslUtils;
@@ -40,19 +43,21 @@ public class GlassfishServerSocketFactory extends JSSE14SocketFactory {
         this.locator = Objects.requireNonNull(locator, "locator");
     }
 
+
     @Override
     public void init() throws IOException {
         sslUtils = locator.getService(SSLUtils.class);
         super.init();
     }
 
+
     @Override
     protected KeyManager[] getKeyManagers(String algorithm, String keyAlias) throws Exception {
         String keystoreFile = (String) attributes.get("keystore");
-        logger.log(Level.FINE, "Keystore file = {0}", keystoreFile);
+        LOG.log(Level.DEBUG, "Keystore file = {0}", keystoreFile);
 
         String keystoreType = (String) attributes.get("keystoreType");
-        logger.log(Level.FINE, "Keystore type = {0}", keystoreType);
+        LOG.log(Level.DEBUG, "Keystore type = {0}", keystoreType);
         KeyManager[] kMgrs = sslUtils.getKeyManagers(algorithm);
         if (keyAlias != null && keyAlias.length() > 0 && kMgrs != null) {
             for (int i = 0; i < kMgrs.length; i++) {
@@ -61,6 +66,7 @@ public class GlassfishServerSocketFactory extends JSSE14SocketFactory {
         }
         return kMgrs;
     }
+
 
     @Override
     protected KeyStore getTrustStore() throws IOException {
